1. combine signin and signup screen into enterScreen 2. create <VerifyVerificationCodeForm /> and use it in VerifyEnterScreen and VerifyIdentityScreen
