+++ src/components/LoginForm/index.js
@@ -1,10 +1,9 @@
-import React, {useEffect, useReducer} from 'react';
+import React, {useEffect} from 'react';
 import {View} from 'react-native';
 import CarrierInfo from 'react-native-carrier-info';
 import {FormattedMessage} from 'react-intl';
 import {useTheme} from 'emotion-theming';
 import {Formik, useFormikContext} from 'formik';
-import useCountDownTimer from '@/hooks/timer';
 
 import Input from '@/components/AppInput';
 import AppButton from '@/components/AppButton';
@@ -12,48 +11,16 @@ import AppText from '@/components/AppText2';
 
 import {
   Container,
-  VerificationContainer,
-  Error,
   PhoneSectionContainer,
   PhonePrefixContainer,
   PhoneContainer,
-  VerificationCodeContainer,
-  sendButtonContainer,
   titleStyle,
   termsStyle,
 } from './style';
 import countryCodeData from '@/constants/countryCode';
 
-const SEND_OTP = 'sendOtp';
-const initialState = {
-  sendCount: 0,
-};
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case SEND_OTP: {
-      return {
-        ...state,
-        sendCount: state.sendCount + 1,
-      };
-    }
-    default:
-      throw new Error();
-  }
-};
-
-const InternalLoginForm = ({submitButtonText, onSendPress}) => {
-  const [state, dispatch] = useReducer(reducer, initialState);
-  const [timeLeft, setCountdownTime] = useCountDownTimer(0);
-  const isTimerStarted = timeLeft > 0;
-
-  const {
-    values,
-    setFieldValue,
-    handleSubmit,
-    errors,
-    isValid,
-  } = useFormikContext();
+const InternalLoginForm = ({submitButtonText, description, theme}) => {
+  const {setFieldValue, handleSubmit, isValid} = useFormikContext();
 
   // get phone prefix
   useEffect(() => {
@@ -62,7 +29,7 @@ const InternalLoginForm = ({submitButtonText, onSendPress}) => {
         const result = await CarrierInfo.isoCountryCode();
         if (result) {
           let dialCode = countryCodeData.find(
-            c => c.code === result.toUpperCase(),
+            (c) => c.code === result.toUpperCase(),
           )?.dial_code;
           if (dialCode) {
             setFieldValue('phonePrefix', dialCode);
@@ -74,14 +41,6 @@ const InternalLoginForm = ({submitButtonText, onSendPress}) => {
     };
     getPhonePrefix();
   }, [setFieldValue]);
-
-  const handleSendPress = async () => {
-    try {
-      setCountdownTime(60);
-      dispatch({type: SEND_OTP});
-      onSendPress(values);
-    } catch (e) {}
-  };
 
   return (
     <View>
@@ -97,44 +56,12 @@ const InternalLoginForm = ({submitButtonText, onSendPress}) => {
           <Input keyboardType="phone-pad" name="phone" />
         </PhoneContainer>
       </PhoneSectionContainer>
-      <VerificationContainer>
-        <VerificationCodeContainer>
-          <Input
-            keyboardType="number-pad"
-            label={
-              <FormattedMessage
-                id="verification_code"
-                defaultMessage="VERIFICATION CODE"
-              />
-            }
-            name="verificationCode"
-          />
-        </VerificationCodeContainer>
-        <AppButton
-          onPress={handleSendPress}
-          text={
-            <>
-              {state.sendCount > 0 ? (
-                <FormattedMessage
-                  id="button.resend_verification_code"
-                  defaultMessage="SEND"
-                />
-              ) : (
-                <FormattedMessage
-                  id="button.send_verification_code"
-                  defaultMessage="RESEND"
-                />
-              )}
-              {isTimerStarted && ' ' + timeLeft}
-            </>
-          }
-          disabled={isTimerStarted || !!errors.phone}
-          variant="filled"
-          sizeVariant="normal"
-          colorVariant="secondary"
-          style={sendButtonContainer}
-        />
-      </VerificationContainer>
+
+      {description && (
+        <AppText variant="caption" style={termsStyle(theme)}>
+          {description}
+        </AppText>
+      )}
       <AppButton
         onPress={handleSubmit}
         text={submitButtonText}
@@ -147,7 +74,7 @@ const InternalLoginForm = ({submitButtonText, onSendPress}) => {
   );
 };
 
-const validate = values => {
+const validate = (values) => {
   const errors = {};
   const dialCodeRegex = /^(\+)(\d{1,3}|\d{1,4})$/;
 
@@ -160,21 +87,6 @@ const validate = values => {
   if (!values.phone) {
     errors.phone = (
       <FormattedMessage id="error.required" defaultMessage="Required" />
-    );
-  }
-
-  if (!values.verificationCode) {
-    errors.verificationCode = (
-      <FormattedMessage id="error.required" defaultMessage="Required" />
-    );
-  }
-
-  if (!!values.verificationCode && values.verificationCode.length !== 6) {
-    errors.verificationCode = (
-      <FormattedMessage
-        id="error.invalid_otp"
-        defaultMessage="OTP is a 6 digit number"
-      />
     );
   }
 
@@ -201,20 +113,16 @@ const LoginForm = ({
         initialValues={{
           phone: '',
           phonePrefix: '',
-          verificationCode: '',
         }}
         onSubmit={onSubmit}
         validate={validate}>
         <InternalLoginForm
           submitButtonText={submitButtonText}
           onSendPress={onSendPress}
+          description={description}
+          theme={theme}
         />
       </Formik>
-      {description && (
-        <AppText variant="caption" style={termsStyle(theme)}>
-          {description}
-        </AppText>
-      )}
     </Container>
   );
 };
