+++ src/intl/zh-HK.json
@@ -8,6 +8,8 @@
   "dark": "夜間模式",
   "light": "日間模式",
   "system_default": "系統默認",
+  "welcome_to_reward_me": "WELCOME to RewardMe",
+  "enter_verification_code": "Enter your verification code",
   
   "valid_until": "Valid until {date}",
   "cashback_type": "Cashback type",
@@ -241,7 +243,8 @@
     "connect_more": "連接更多",
     "sure": "好吧!",
     "continue": "繼續",
-    "upgrade": "升級"
+    "upgrade": "升級",
+    "verify": "驗證"
   },
 
   "error": {
