--- src/screens/VerifyIdentityScreen/style.js
@@ -1,17 +1,0 @@
-import {css} from '@emotion/native';
-
-export const detailStyle = theme => css`
-  color: ${theme.colors.textOnBackground.mediumEmphasis};
-`;
-
-export const container = css`
-  padding-horizontal: 24px;
-`;
-
-export const formContainer = css`
-  margin-vertical: 24px;
-`;
-
-export const button = css`
-  margin-top: 24px;
-`;
